♻️refactor : 컴포넌트 정리
