+++ src/routes/Home.js
@@ -1,21 +1,11 @@
-import { useEffect, useState, useRef } from 'react'
-import { db, storage } from 'myBase'
-import {
-  addDoc,
-  collection,
-  serverTimestamp,
-  query,
-  onSnapshot,
-  orderBy,
-} from 'firebase/firestore'
-import { v4 } from 'uuid'
-import { ref, getDownloadURL, uploadString } from '@firebase/storage'
-import Sweet from 'components/Sweet'
+import { useEffect, useState } from 'react'
+import { db } from 'myBase'
+import { collection, query, onSnapshot, orderBy } from 'firebase/firestore'
+import EditSweet from 'components/EditSweet'
+import WriteSweets from 'components/WriteSweets'
 
 export default function Home({ userObj }) {
-  const [sweet, setSweet] = useState('')
   const [sweets, setSweets] = useState([])
-  const [attachment, setAttachment] = useState('')
 
   useEffect(() => {
     const q = query(collection(db, 'sweets'), orderBy('createdAt', 'desc'))
@@ -28,89 +18,14 @@ export default function Home({ userObj }) {
     })
   }, [])
 
-  const onSubmitSweet = async (e) => {
-    e.preventDefault()
-    let attachmentUrl = ''
-    try {
-      if (attachment !== '') {
-        const attachmentRef = ref(storage, `${userObj.uid}/${v4()}`)
-        const attachmentUpload = await uploadString(
-          attachmentRef,
-          attachment,
-          'data_url',
-        )
-        attachmentUrl = await getDownloadURL(attachmentUpload.ref)
-      }
-      const doc = await addDoc(collection(db, 'sweets'), {
-        text: sweet,
-        createdAt: serverTimestamp(),
-        creator: userObj.uid,
-        attachmentUrl,
-      })
-      console.log(doc)
-    } catch (err) {
-      console.log(err.message)
-    }
-    setSweet('')
-    onClickClear()
-  }
-
-  const onChangeSweet = ({ target: { value } }) => {
-    setSweet(value)
-  }
-
-  const onFileChange = (e) => {
-    const {
-      target: { files },
-    } = e
-    const fileInfo = files[0]
-    const reader = new FileReader()
-    reader.onloadend = (finishedEvent) => {
-      const {
-        currentTarget: { result },
-      } = finishedEvent
-      setAttachment(result)
-    }
-    reader.readAsDataURL(fileInfo)
-  }
-  const fileInput = useRef()
-  const onClickClear = () => {
-    fileInput.current.value = ''
-    setAttachment('')
-  }
-
   return (
     <>
-      <form onSubmit={onSubmitSweet}>
-        <input
-          type='text'
-          placeholder="What's on your mind?"
-          maxLength={140}
-          value={sweet}
-          onChange={onChangeSweet}
-        />
-        <input type='submit' value='Sweet' />
-        <input
-          type='file'
-          accept='image/*'
-          onChange={onFileChange}
-          ref={fileInput}
-        />
-        {attachment && (
-          <div>
-            <img
-              src={attachment}
-              width='120px'
-              height='120px'
-              alt='thumbnail'
-            />
-            <button onClick={onClickClear}>Delete</button>
-          </div>
-        )}
-      </form>
+      <div>
+        <WriteSweets userObj={userObj} />
+      </div>
       <div>
         {sweets.map(({ id, text, creator, attachmentUrl }) => (
-          <Sweet
+          <EditSweet
             key={id}
             id={id}
             txt={text}
